succesfully moved server files, database, configs and tests  into server folder
