+++ server/db/models/vote.js
